Fix converter deprecation warning tests and move to test_connection.py
The DeprecationWarning tests failed because the process-wide dedup set
(_BACKWARD_COMPAT_WARNED) was consumed by earlier tests accessing the
properties via helper functions, preventing the warning from firing a
second time. Fixed by clearing the dedup slot before each deprecation
test. Moved all converter tests into test_connection.py for cohesion.
