+++ python/tests/integ/test_connection.py
@@ -3,6 +3,7 @@ Integration tests for PEP 249 Connection objects.
 """
 
 import uuid
+import warnings
 
 from io import StringIO
 from typing import Any
@@ -13,6 +14,12 @@ import pytest
 from snowflake.connector.constants import QueryStatus
 from snowflake.connector.cursor import DictCursor
 from snowflake.connector.errors import DatabaseError, InterfaceError, ProgrammingError
+
+
+with warnings.catch_warnings():
+    warnings.simplefilter("ignore", DeprecationWarning)
+    from snowflake.connector.converter import SnowflakeConverter
+    from snowflake.connector.converter_null import SnowflakeNoConverterToPython
 
 
 # These tests heavily mutate the connection (close, autocommit, commit, rollback,
@@ -615,3 +622,106 @@ class TestIsValid:
         conn = connection_factory()
         conn.close()
         assert conn.is_valid() is False
+
+
+def _get_converter_class(connection):
+    """Read converter_class while suppressing the DeprecationWarning."""
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        return connection.converter_class
+
+
+def _get_converter(connection):
+    """Read converter while suppressing the DeprecationWarning."""
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        return connection.converter
+
+
+class TestConverterClassProperty:
+    """Integration tests for Connection.converter_class."""
+
+    def test_converter_class_returns_a_type(self, connection):
+        """converter_class should return a class (type), not an instance."""
+        assert isinstance(_get_converter_class(connection), type)
+
+    def test_converter_class_is_snowflake_converter_or_subclass(self, connection):
+        """The default converter_class should be SnowflakeConverter or a subclass of it."""
+        assert issubclass(_get_converter_class(connection), SnowflakeConverter)
+
+    def test_converter_class_accessible_on_open_connection(self, connection):
+        """converter_class should be accessible on an open connection without error."""
+        cls = _get_converter_class(connection)
+        assert cls is not None
+
+    def test_converter_class_on_closed_connection(self, connection_factory):
+        """converter_class should remain accessible after the connection is closed."""
+        conn = connection_factory()
+        cls_before = _get_converter_class(conn)
+        conn.close()
+        cls_after = _get_converter_class(conn)
+        assert cls_before is cls_after
+
+    def test_converter_class_custom_subclass(self, connection_factory):
+        """Passing a custom converter_class subclass should be accepted."""
+
+        class MyConverter(SnowflakeConverter):
+            pass
+
+        with connection_factory(converter_class=MyConverter) as conn:
+            assert _get_converter_class(conn) is MyConverter
+
+    def test_no_converter_to_python_is_subclass_of_snowflake_converter(self):
+        """SnowflakeNoConverterToPython should be a subclass of SnowflakeConverter."""
+        assert issubclass(SnowflakeNoConverterToPython, SnowflakeConverter)
+
+    def test_no_converter_to_python_as_converter_class(self, connection_factory):
+        """Passing SnowflakeNoConverterToPython as converter_class should be accepted."""
+        with connection_factory(converter_class=SnowflakeNoConverterToPython) as conn:
+            assert _get_converter_class(conn) is SnowflakeNoConverterToPython
+
+    @pytest.mark.skip_reference(reason="DeprecationWarning is universal-driver only")
+    def test_converter_class_emits_deprecation_warning(self, connection):
+        """Accessing converter_class should emit a DeprecationWarning."""
+        from snowflake.connector._internal.backward_compatibility import _BACKWARD_COMPAT_WARNED
+
+        # Clear the process-wide dedup slot so the warning fires even if an
+        # earlier test already accessed the property (which silently consumed
+        # the one-shot slot via _emit_backward_compatibility_warning).
+        _BACKWARD_COMPAT_WARNED.discard(("snowflake.connector.connection", "Connection.converter_class"))
+        with warnings.catch_warnings(record=True) as caught:
+            warnings.simplefilter("always")
+            _ = connection.converter_class
+        deprecations = [w for w in caught if issubclass(w.category, DeprecationWarning)]
+        assert len(deprecations) == 1
+        assert "converter_class" in str(deprecations[0].message)
+
+
+class TestConverterProperty:
+    """Integration tests for Connection.converter (instance)."""
+
+    def test_converter_is_instance_of_converter_class(self, connection):
+        """converter should be an instance of the connection's converter_class."""
+        assert isinstance(_get_converter(connection), _get_converter_class(connection))
+
+    def test_converter_default_is_snowflake_converter_instance(self, connection):
+        """With default settings, converter should be a SnowflakeConverter instance."""
+        assert isinstance(_get_converter(connection), SnowflakeConverter)
+
+    def test_converter_with_custom_class(self, connection_factory):
+        """When converter_class is overridden, converter should be an instance of that class."""
+        with connection_factory(converter_class=SnowflakeNoConverterToPython) as conn:
+            assert isinstance(_get_converter(conn), SnowflakeNoConverterToPython)
+
+    @pytest.mark.skip_reference(reason="DeprecationWarning is universal-driver only")
+    def test_converter_emits_deprecation_warning(self, connection):
+        """Accessing converter should emit a DeprecationWarning."""
+        from snowflake.connector._internal.backward_compatibility import _BACKWARD_COMPAT_WARNED
+
+        _BACKWARD_COMPAT_WARNED.discard(("snowflake.connector.connection", "Connection.converter"))
+        with warnings.catch_warnings(record=True) as caught:
+            warnings.simplefilter("always")
+            _ = connection.converter
+        deprecations = [w for w in caught if issubclass(w.category, DeprecationWarning)]
+        assert len(deprecations) == 1
+        assert "converter" in str(deprecations[0].message)
